refactor(LC017): split whole-entity projection fixer helpers
- separate fix-context analysis and projection rewrite logic
- keep fixer behavior unchanged while reducing local complexity
